Add an index structure for counter event sets
This patch adds a data structure capable of indexing a counter event
set to speed up operations that must determine the minimum and maximum
value or slope within an interval of time. The index is an n-ary
search tree whose nodes accumulate the minimum and maximum values /
slopes for their children.