BUGFIX: counter_event_set_index: Use slope of the right sample instead of the left
The slope between two samples A and B is stored in the data structure
associated to B. Hence, when querying the slope for any point T
between the two samples the slope associated to B should be
returned. However, the counter_event_set_index used the slope of A,
leading to a shift of one sample in the graphical representation. This
patch fixes the code of
counter_event_set_index_node_min_max_slope_leaf and
counter_event_set_index_node_min_max_slope_internal and updates the
tests in tests/counter_event_set_index.test.c.