BUGFIX: event_set_get_remote_local_numa_bytes_in_interval: Consider partial overlaps
Under certain conditions, a task starting outside the query interval
and ending within the interval is not considered by
event_set_get_remote_local_numa_bytes_in_interval. This modification
checks if such a task exists and initializes the pointer to the first
task to be taken into account accordingly.