Add support for filter expressions to the filter infrastructure
A filter expression may be composed of several, whitespace-separated
sub-expression, where each sub-expression has the following form:

  - A mode character (+, - or !). A plus sign indicates that the
    sub-expression adds something to the filter, the minus sign
    indicates removal from the filter while the exclamation mark sets
    the filter to a certain value. For certain filter functions the
    mode character is not required (e.g., for tasklen; see below).

  - A filter function (e.g., cpu, task, tasklen, event) that indicates
     which part of the filter the sub-expression applies to.

  - A function-specific argument in parantheses (e.g., the range of
    CPUs to be added to the filter, the interval of task durations for
    the tasklen function, etc.)

Depending on the function an argument can either be a single integer
with an optionaml unit prefix (e.g., 5, 123, 19K or 189M), a range of
integers in brackets with optional unit prefixes (e.g., [1-10],
[500-55K], [100M-1G]), a string in quotes (e.g., "foo", "a string with
\"quotes\"") or a POSIX extended regular expression in quotes prefixed
by the character r (e.g., r"^init", r"[a-zA-Z]*").

Examples:

  - "+cpu(5)" adds cpu 5 to f.

  - "-cpu(5)" removes cpu 5 from f.

  - "!cpu(5)" clears the set of CPUs of f before adding cpu 5.

  - "tasklen([5K-100M])" configures the filter to include only task
    instances whose duration is between 5000 and 100000000 cycles.

  - "+task(\"init\")" adds all tasks to f whose associated symbol
    is equal to "init".

  - "!task(\"^init.*\")" configures f to include only tasks whose
    associated symbol starts with the string "init".

  - "!event(tdestroy)" configures f to include only task
    destruction events.

A filter function does not affect the filter for any other aspect of
the filter. For example, limiting the filter to CPU 5 using "!cpu(5)"
neither affects the task types included in the filter, nor does it
affect the set of single events of the filter.