Add generic function for drawing CPU lanes on the time line
For each time line mode (states, heatmap, type map, numa maps, etc.)
there is a dedicated function for rendering. The new function
gtk_trace_paint_generic factors the code shared by these functions and
provides a generic way of rendering the timeline. For each horizontal
pixel of each visible lane it calls a callback function of the form:

  int pxfun(void* data, int cpu, uint64_t start, uint64_t end,
            double* r, double* g, double* b);

where data is the address of data provided by the caller of
gtk_trace_paint_generic, cpu_idx is the index of the CPU whose lane is
rendered, start and end are the start and end timestamps of the
associated to the interval of the current pixel, and r, g and b are
output parameters for the color components.

The return value of the callback function must indicate whether the
color should be taken into account. For a return value different from
0 gtk_trace_paint_generic will render the pixel, while a return value
of 0 causes the pixel to be rendered transparently.