Makefile.am: Add targets for valgrind remote debugging
Valgrind can be instructed to suspend execution on an error and to
wait for a debugger to connect. This can be done for any unit test
binary with the new target valgrind-gdb-%. When valgrind waits for a
debugger, the target connect-valgrind-gdb-% can then be used to start
gdb and to connect to valgrind. Arguments to be passed to gdb can be
specified in the environment variable GDB_FLAGS.