Libaftermath-core: Object notation: Add support for selection expressions
An expression can be composed of a concatenation separated with points of the
following basic expressions:

  - A member expression (only valid on groups), composed of the name of the
    group member

  - A list index expression (only valid on lists), represented by an integer in
    brackets

Example:

  person {
    name: "foo",
    age: 30,
    favorite_meals: [ "pizza", "spaghetti", "hamburger", "chicken" ]
  }

To select the name or age of the person, one would evaluate the expression
"name" or "age", respectively. To get the first favorite meal, one would
evaluate "favorite_meals[0]", for the second favorite meal "favorite_meal[1]",
and so on.