libaftermath-core: Do not include sys/stat.h breaking builds with recent C++ compilers
Aftermath build fails on recent system when the C header sys/stat.h is
included in C++ code base. Types like "__u64" generate an error for the
C++ compiler.

This patch uses open and lseek in place of stat to obtain the file size,
getting rid of the header requirement.