BUGFIX: Aftermath QT GUI: Fix segfault for ill-formed node positions in DFG
When loading a DFG with an ill-formed list of node positions (e.g., with an
entry having three coordindates instead of two), the QT GUI catches the error,
but terminates with a segmentation fault.

The error is due to a double destruction of the DFG created from the DFG object
notation: once in a catch block in AftermathSession::loadDFG() and once again in
the destructor of AftermathSession.

By setting the session's DFG only after validation of the list of node positions
after the catch block, the destructor of AftermathSession does not attempt to
destroy the DFG a second time.