Libaftermath-render: Add support for rendering of discrete events
This adds a new lane render layer meta type for discrete events (i.e., events
that occur at a timestamp, but that do not have a duration). Similarly to the
interval render layer type, this is not a concrete type, but a meta type that
provides common infrastructure to implement actual render layer types for
discrete events.

The actual type must provide a custom rendering function, which is invoked by
the discrete event rendering layer for each pixel for which at least one event
was found in the corresponding interval.