Libaftermath-core: statistics/interval.h: Remove include of state_event_array.h
The code in statistics/interval.h does not use any declaration from
state_event_array.h, but from files that are included by that header file.

This patch replaces the inclusion of statistics/interval.h with include
directives of the header files which provide the actual declarations that are
needed.