Libaftermath-core: Binary search macros: Use T instead of NEEDLE_T for iterator
The pointer to the current element in the code generated by the macros for
binary search functions is of type NEEDLE_T (type of the query
element). However, the elements of the array are of type T. This code has worked
so far, as all invocations of the macros used the same type for T and NEEDLE_T,
but it fails if the types are different.

This patch uses T instead of NEEDLE_T for the current element and enables
invocations of the macros, in which T and NEEDLE_T are different.