Libaftermath-core: LoadFunction.tpl.c: Only generate destruction label if needed
For on-disk types with a destructor that do not have an assertion or process
function, the label in the load function for destruction of the type is
generated, but never used. This results in a compilation error with the default
compiler options -Wall and -Werror.

Only generate the label if its is actually used.