Libaftermath-render: State event renderer: Unconditionally initialize statistics
Initialization of the maximum index for statistics is currently skipped in
state_renderer_trace_changed() if the state event description array of the newly
set trace is empty. However, the interval layer rendering infrastructure is
unaware of descriptors related to the rendered events and thus unconditionally
invokes functions gathering statistics, which in turn rely on information about
the maximum index. This leads to reads of uninitialized memory locations under
certain conditions.

This patch fixes the above-mentioned problem by unconditionally setting the
maximum index.